+++ src/hooked/__main__.py
@@ -27,7 +27,7 @@
 
 from __future__ import annotations
 
-from hooked.main import main
+from hooked.hooked import main
 
 if __name__ == "__main__":
     raise SystemExit(main())
